chore: move tests to tests/
Move tests to tests/ which is a bit cleaner but also works better with
setuptools. Update Makefile and README.md for new tests/.

In addition, with use of venv, explicitly setting PYTHONPATH is no
longer needed, so update Makefile and README.md accordingly.
