+++ Makefile
@@ -11,14 +11,17 @@ check-deb-deps:
 check-deps: check-deb-deps
 
 test:
-	PYTHONPATH=./ ./cvelib/run-tests
+	if test -z "$(VIRTUAL_ENV)" ; then \
+		echo "WARN: not running in venv. Did you forget to '. ./.venv/bin/activate'? Proceeding anyway..." ; \
+	fi
+	./tests/run-tests
 
 syntax-check: clean
-	./cvelib/run-flake8
-	./cvelib/run-pylint
+	./tests/run-flake8
+	./tests/run-pylint
 
 style-check: clean
-	./cvelib/run-black
+	./tests/run-black
 
 # require woke to be installed in CI but not one local system
 inclusivity-check: clean
@@ -34,11 +37,11 @@ inclusivity-check: clean
 check: check-deps test inclusivity-check syntax-check style-check
 
 coverage: check-deb-deps
-	PYTHONPATH=./ python3 -m coverage run ./cvelib/run-tests
+	python3 -m coverage run ./tests/run-tests
 
 coverage-report:
 	python3 -m coverage report --show-missing --omit="*/dist-packages/*"
 
 clean:
-	rm -rf ./bin/__pycache__ ./cvelib/__pycache__
+	rm -rf ./bin/__pycache__ ./cvelib/__pycache__ ./tests/__pycache__
 	rm -rf .coverage
