+++ tests/test_pkg.py
